Add Copy trait to Iter impl blocks
In previous pattern, Iter didn't not require that T::Value is Copy.
The impl block for Iterator on Iter did require that T::Value is Copy.
This allowed for an iterator to be created over an entry that could not
call the next method, and therefore could not use any of the Iterator
trait methods. Copy trait added to Iter creation so that calling .iter()
on such entires would fail to build.