Add test for Tendermint::init()
This creates a temporary directory for the Tendermint instance to be
created into and then deletes the directory and its contained files
when the TempDir is dropped. Could also go a bit deeper in this log to
flatten the files in config and data into the HashSet to ensure that
everything was created correctly.