Add Deque::retain implementation
Given that the inner Map may be attached to a Store at any given time,
unretained values bubbled to the back of the Deque must be popped off
rather than removed by updating the tail index as in VecDeque. Because
of this, Deque::retain is Θ(n), but retains the ordering of elements.